build(Project barebones with pygame window for game): Made the framework for the project where a black window is made, with a simple function to run game.
This project will use the pygame module, and with it I created a class for the main logic of the game. This results in a black screen shown where elements wil sit on. To control the framerate of the game,  a clock was used to make the game run at a constant rate on all systems.