build(Wizard class,  refactored settings, add to run_game logic): Made class for wizard player to be displayed as small blue box, refactored settings and added to main game logic.
Placed settings into it's own file for easier to used and changed for different game elements. Then created a class for the wizard player where it will be displayed as a small blue cube at the bottom of the screen.  Then once the wizard class was made, changes were made to the main file to allow it to be draw the wizard every frame in `run_game()`.