feat(Horizonal movement with flags): Added movement flags for the wizard to move continuously on the X-axis.
Movement flags were chosen to be used because they let the wizard keep moving smoothly while a key is held down, rather than only reacting at the moment the key is pressed. When a key is pressed, the flag turns on, and when it’s released, the flag turns off. This creates much more natural and responsive controls for the player. However an issue right now is how this also let's the wizard disappear off screen.