feat(Fix boundary, and new feature): Fixed boundary issue in `update()`, and added ability to move on the Y-axis.
The issue for the wizard going off screen was fixed by in `update()`, where it finds the edge of the screen and position of the edge of the face facing the boundary.  Then another setting to increase moving speed was added.  To allow the wizard to move in the y-axis, the up and down arrow keys were registered with flags attached to them as well.  Now the wizard can zip all over the  screen. It could be possilbe to code for diagonal movemnt but the game works witihout it.