feat(Including the spells element for this game.): Registered spacebar key to a helper function to fire spells.
In order to treat all spells like a group,  in pygame they were handled using the sprite.Group class.  Then to fire the spells, the spacebar was keyed to be the trigger. It then  calls the helper mthod  `_fire_spell` to make a new instance of a spell and then added it to the group.