feat(Collisions): Added a way to handle collisions, and changed _create_enemies to now make more  monsters.
Now when you hit a monster, both it and the spell that hit it will disappear. Then to keep the amount of monsters at a constant amount, as soon as one dies it will check against the setting limit and then make more.  I may make a setting later to change it to a delay incase it becomes too hard having them instantly respawn.