feat(Monster Movement): Built function to use the Math module to calculate distance to wizard and updated main file to allow movement.
In the monser.py file, `update()` function was made to find the distance to the wizard from the position of a monster. I thought of how  since this game has an x and y value, and how the wizard does too, then it's basically just like the Pythagorean theory. Then to make the monster move in a smooth way, I made it so it moves by the monster speed setting. Right now it works, but I do notice the monsters tend to cluster in a small group after a while which makes running away from them too easy at this point. I will need to fix that to make it acutally much harder.