feat(Scoring): Add scoreboard class, and display it on-screen in green text on the top right corner.
Made a second class in game_stats to make a scoring system. This class has the settings to place the score in the top right corner, and then draw it on-screen. The main file was then updated to  make an instance and then draws a blank 0 for now  to show it works.