feat(Wizard Lives): Add feature to display  the number of lives the wizard has on the top left corner.
To show the lives that the player has  I used the code to make the score display. This works for now, but I may change it later on to condense it into fewer functions.  In the main file, it only needed a small change in _wizard_hit() to show the lives on the screen.